Move all local integrations in council-frontend to src/integrations instead of prefixing with elf-
