+++ apps/council-frontend/src/ui/contracts/useChangeDelegation.tsx
@@ -7,7 +7,7 @@ import { makeSmartContractReadCallQueryKey } from "@elementfi/react-query-typech
 import { useSmartContractTransaction } from "@elementfi/react-query-typechain/src/hooks/useSmartContractTransaction/useSmartContractTransaction";
 import { queryClient } from "src/queryClient";
 import { useRef } from "react";
-import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/elf-etherscan/domain";
+import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/integrations/etherscan";
 import { t, jt } from "ttag";
 import ExternalLink from "src/ui/base/ExternalLink/ExternalLink";
