+++ apps/council-frontend/src/ui/featureFlag/useFeatureFlag.ts
@@ -1,5 +1,5 @@
 import { useLocalStorage } from "react-use";
-import { FeatureFlag } from "src/featureFlag";
+import { FeatureFlag } from "src/featureFlags";
 
 export function useFeatureFlag(flagKey: FeatureFlag): boolean {
   const [flag] = useLocalStorage(flagKey, false);
