+++ apps/council-frontend/src/ui/proposals/ProposalsDetailsCard/GSCProposalsDetailsCard.tsx
@@ -15,7 +15,7 @@ import { isNumber } from "lodash";
 import { jt, t } from "ttag";
 
 import { getIsVotingOpen } from "src/elf-council-proposals";
-import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/elf-etherscan/domain";
+import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/integrations/etherscan";
 import { defaultProvider } from "src/providers/providers";
 import ElementUrl from "src/urls";
 import Button from "src/ui/base/Button/Button";
