+++ apps/council-frontend/src/ui/proposals/ProposalsDetailsCard/ProposalDetailsCard.tsx
@@ -15,7 +15,7 @@ import React, {
 import toast from "react-hot-toast";
 import ReactMarkdown from "react-markdown";
 import { getIsVotingOpen } from "src/elf-council-proposals";
-import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/elf-etherscan/domain";
+import { ETHERSCAN_TRANSACTION_DOMAIN } from "src/integrations/etherscan";
 import { VotingPower } from "src/proposals/VotingPower";
 import ElementUrl from "src/urls";
 import { BalanceWithLabel } from "src/ui/base/BalanceWithLabel/BalanceWithLabel";
